+++ app/src/main/java/com/fullstack/frontend/onboard/login/LoginFragment.java
@@ -1,13 +1,16 @@
 package com.fullstack.frontend.onboard.login;
 
 import android.content.Context;
+import android.os.Build;
 import android.os.Bundle;
+import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.Toast;
 
 import androidx.annotation.NonNull;
+import androidx.annotation.RequiresApi;
 import androidx.lifecycle.LiveData;
 import androidx.lifecycle.ViewModel;
 import androidx.lifecycle.ViewModelProvider;
@@ -15,13 +18,24 @@ import androidx.navigation.NavController;
 import androidx.navigation.Navigation;
 
 import com.fullstack.frontend.R;
+import com.fullstack.frontend.Retro.ApiClient;
+import com.fullstack.frontend.Retro.ApiInterface;
+import com.fullstack.frontend.Retro.BaseResponse;
 import com.fullstack.frontend.Retro.OnBoardingResponse;
+import com.fullstack.frontend.Retro.TokenRequest;
 import com.fullstack.frontend.base.BaseFragment;
 import com.fullstack.frontend.base.RemoteRequestListener;
 import com.fullstack.frontend.config.UserInfo;
 import com.fullstack.frontend.databinding.FragmentLoginBinding;
 import com.fullstack.frontend.databinding.FragmentRegisterBinding;
 import com.fullstack.frontend.util.Util;
+import com.google.android.gms.tasks.OnSuccessListener;
+import com.google.firebase.iid.FirebaseInstanceId;
+import com.google.firebase.iid.InstanceIdResult;
+
+import retrofit2.Call;
+import retrofit2.Callback;
+import retrofit2.Response;
 
 
 public class LoginFragment extends BaseFragment<LoginViewModel, LoginModel> implements RemoteRequestListener {
@@ -97,8 +111,54 @@ public class LoginFragment extends BaseFragment<LoginViewModel, LoginModel> impl
     public void onSuccess(LiveData<OnBoardingResponse> loginResponse) {
         loginResponse.observe(this, it -> {
             UserInfo.setUser_id(it.getId());
+            Log.d("11", "set user id: " + UserInfo.getUser_id());
             Toast.makeText(getActivity(), "Login Successfully!", Toast.LENGTH_SHORT).show();
+
+
+            FirebaseInstanceId.getInstance().getInstanceId().addOnSuccessListener(new OnSuccessListener<InstanceIdResult>() {
+                @Override
+                public void onSuccess(InstanceIdResult instanceIdResult) {
+
+
+                    Log.d("token", instanceIdResult.getToken());
+
+                    ApiInterface apiService = ApiClient.getClient().create(ApiInterface.class);
+                    //Log.d(TAG, "user id:" + UserInfo.getUser_id());
+                    TokenRequest request = new TokenRequest(UserInfo.getUser_id(), instanceIdResult.getToken());
+                    Log.d("11", "user id: " + UserInfo.getUser_id() + "token: " + instanceIdResult.getToken());
+                    Call<BaseResponse> postToken = apiService.postToken(request);
+                    postToken.enqueue(new Callback<BaseResponse>() {
+                        @RequiresApi(api = Build.VERSION_CODES.N)
+                        @Override
+                        public void onResponse(Call<BaseResponse> call, Response<BaseResponse> response) {
+                            if (response.isSuccessful()){
+
+                                Log.d("FirebaseService", "Send token: " + response.code());
+                            }
+                            if (response.body() != null) {
+                                Log.d("FirebaseService", "Send token: " + "Success");
+
+
+                            }
+                        }
+
+                        @Override
+                        public void onFailure(Call<BaseResponse> call, Throwable t) {
+
+                            Log.d("FirebaseService", "Send token failed");
+
+
+                        }
+                    });
+                }
+            });
+
+
+
             Navigation.findNavController(getView()).navigate(R.id.nav_home);
+
+
+
         });
     }
